+++ assets/styles/styles.css
@@ -66,7 +66,8 @@
   --bs-dark-border-subtle: #adb5bd;
   --bs-white-rgb: 255, 255, 255;
   --bs-black-rgb: 0, 0, 0;
-  --bs-font-sans-serif: Helvetica, Arial, sans-serif;
+  /* --bs-font-sans-serif: Helvetica, Arial, sans-serif; */
+  --bs-font-sans-serif: system-ui, -apple-system, "Segoe UI", Roboto, "Helvetica Neue", "Noto Sans", "Liberation Sans", Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji";
   --bs-font-monospace: SFMono-Regular, Menlo, Monaco, Consolas, "Liberation Mono", "Courier New", monospace;
   --bs-gradient: linear-gradient(180deg, rgba(255, 255, 255, 0.15), rgba(255, 255, 255, 0));
   --bs-body-font-family: var(--bs-font-sans-serif);
@@ -133,6 +134,7 @@
   }
 }
 
+
 body {
   margin: 0;
   font-family: var(--bs-body-font-family);
@@ -145,8 +147,6 @@ body {
   -webkit-text-size-adjust: 100%;
   -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
 }
-
-
 
 hr {
   margin: 1rem 0;
@@ -12229,6 +12229,8 @@ fieldset:disabled .btn {
     display: none !important;
   }
 }
+
+
 .btn-give {
   --bs-btn-font-weight: 700;
   --bs-btn-color: var(--bs-white);
@@ -12256,6 +12258,9 @@ fieldset:disabled .btn {
 header {
   -webkit-box-shadow: 0 4px 3px 0 rgba(0, 0, 0, 0.1);
   box-shadow: 0 4px 3px 0 rgba(0, 0, 0, 0.1);
+
+  
+  
 }
 
 /*** ROW-1 HEIGHTS ***/
@@ -12279,6 +12284,7 @@ header .row-1 {
     opacity: 1;
   }
 }
+
 #kcmh .main-nav .btn-close {
   display: none;
 }
@@ -12302,6 +12308,8 @@ header .row-1 {
     padding-top: 2px;
   }
 }
+
+
 .navbar-brand img {
   width: 171px;
   height: 32px;
@@ -12313,6 +12321,8 @@ header .row-1 {
     height: 57px;
   }
 }
+
+
 @media screen and (min-width: 992px) {
   #kcmh {
     padding: 0 20px;
@@ -12606,9 +12616,13 @@ ul.nav.kcmh-nav-member {
 
 @media (min-width: 992px) {
   .region-map-icon {
-    margin-right: 8px;
+    /* margin-right: 8px; */
+    margin-right: 4px
+    
   }
 }
+
+
 @media (min-width: 1200px) {
   .region-map-icon {
     margin-right: 12px;
