--- components/PageHeader-initial.vue
@@ -1,36 +1,0 @@
-
-<template>
-    
-  <header class="container" id="page-header">
-    1. Page Header
-  </header>
-
-</template>
-
-<script setup>
-
-</script>
-
-<style scoped>
-
- 
-  
-#page-header {
-  padding-top: 16px;
-  padding-bottom: 16px;
-  height: 60px;
-  
-  text-align: center;
-  
-}
-
-  
-@media (min-width: 768px) {
-  #page-header {
-    height: 163px;
-
-  }
-
-}
-  </style>
-
